pwmgr/requirements.txt
cffi==1.17.1
cryptography==45.0.2
pychalk==2.0.1
pycparser==2.22
pyperclip==1.9.0
six==1.17.0
